requirements.txt
python-dotenv[cli]
langchain
huggingface_hub
streamlit
